requirements.txt
# Core dependencies
pathlib>=1.0.1
typing-extensions>=4.1.1
aiohttp>=3.8.5
python-dotenv>=1.0.0

# Code formatting and linting
black>=22.3.0
isort>=5.10.1
mypy>=0.942
pylint>=2.13.0
flake8>=5.0.0

# Testing
pytest>=7.1.1
pytest-asyncio>=0.18.3

# Environment detection
platformdirs>=2.5.1

# LLM integration
openai>=1.3.0

# Code transformation
libcst>=1.0.1
networkx>=3.0
pydeps>=1.10.0  # For dependency visualization
GitPython>=3.1.30  # For git-like versioning of changes